fix(main): cast to fp32

+++ src/main.py
@@ -56,8 +56,8 @@ def jitless_step(while_ctx_dict: typing.Dict[str, typing.Any]) -> typing.Dict[st
     wctx.data = jnp.stack([data[:, :, :-1], data[:, :, 1:]], 1)
 
     out = WhileTrainContext(loop(train_step, wctx.serialize(), steps, training.device_unroll))
-    out.loss = data.min()
-    out.top_loss = data.max()
+    out.loss = data.min().astype(jnp.float32)
+    out.top_loss = data.max().astype(jnp.float32)
     return out.serialize()
 
 def get_parameters(ctx: Context, inp: jnp.ndarray):
